reduce number of redundant test cases; add defensive sorting of usages
